ADD : 환자 정보, 완치자 정보 API 구현

+++ src/main/java/com/example/toothfairy/web/dto/CuredInfoDto.java
@@ -16,7 +16,7 @@ import java.sql.Date;
 public class CuredInfoDto {
     private int age;
 
-    private Long totalTreatmentDate; // 총 치료 기간 (milliseconds)
+    private int totalTreatmentDate; // 총 치료 기간 (milliseconds)
     private Long totalWearTime;      // 총 착용 시간 (milliseconds)
 
     public static CuredInfoDto createDto(CuredInfo entity){
